fix all errors with lint-fix and ran a local build to ensure everything works now (will handle merge conflicts after role is completed)
