+++ app/api/auth/[...nextauth]/route.ts
@@ -1,74 +1,6 @@
-import { PrismaAdapter } from "@next-auth/prisma-adapter";
-import type { NextAuthOptions, Session as _NextAuthSession } from "next-auth";
-import NextAuth, { getServerSession } from "next-auth";
-import GoogleProvider from "next-auth/providers/google";
-import prisma from "../../../../lib/prisma";
+import NextAuth from "next-auth";
+import { authOptions } from "@/lib/auth";
 
-export const authOptions: NextAuthOptions = {
-    session: {
-        strategy: "jwt",
-    },
-    adapter: PrismaAdapter(prisma),
-    providers: [
-        GoogleProvider({
-            clientId: process.env.GOOGLE_ID ?? "",
-            clientSecret: process.env.GOOGLE_SECRET ?? "",
-            profile(profile: {
-                sub: string;
-                given_name?: string;
-                family_name?: string;
-                email: string;
-            }) {
-                return {
-                    id: profile.sub,
-                    firstName: profile.given_name ?? "",
-                    lastName: profile.family_name ?? "",
-                    email: profile.email,
-                };
-            },
-        }),
-    ],
-    callbacks: {
-        async jwt({ token, user }) {
-            if (user) {
-                token.id = user.id;
-                const typedUser = user as {
-                    id: string;
-                    email?: string;
-                    firstName?: string;
-                    lastName?: string;
-                    finishedOnboarding?: boolean;
-                };
-                token.firstName = typedUser.firstName;
-                token.lastName = typedUser.lastName;
-                token.firstTimeUser = !typedUser.finishedOnboarding;
-            } else {
-                // On subsequent token refreshes (e.g. via update())
-                // Re-fetch the latest user record from the database.
-                const dbUser = await prisma.user.findUnique({
-                    where: { id: token.id },
-                    select: { finishedOnboarding: true },
-                });
-                if (dbUser) {
-                    token.firstTimeUser = !dbUser.finishedOnboarding;
-                }
-            }
-            return token;
-        },
-        async session({ session, token }) {
-            if (session.user) {
-                session.user.id = token.id;
-                session.user.firstName = token.firstName ?? "";
-                session.user.lastName = token.lastName ?? "";
-                (session as any).firstTimeUser = token.firstTimeUser;
-            }
-            return session;
-        },
-    },
-    secret: process.env.NEXTAUTH_SECRET,
-};
+const handler = NextAuth(authOptions) as (req: Request) => Promise<Response>;
 
-const handler = NextAuth(authOptions);
-
-export const getServerAuthSession = () => getServerSession(authOptions);
 export { handler as GET, handler as POST };
