+++ app/signup/name/page.tsx
@@ -165,72 +165,3 @@ export default function Name() {
         </main>
     );
 }
-//     return (
-//         <main className="min-h-screen bg-gray-50">
-//             {/* Content section with input fields */}
-//             <section className="flex justify-center pt-20">
-//                 <div className="flex flex-col space-y-8 content-center justify-center gap-1">
-//                     <p className="text-center text-2xl">Enter your name:</p>
-
-//                     {/* First Name Input */}
-//                     <label className="outline outline-1 pl-8 pr-20 py-3 rounded-lg bg-white">
-//                         <input
-//                             type="text"
-//                             value={firstName}
-//                             onChange={(e) => {
-//                                 setFirstName(e.target.value);
-//                             }}
-//                             className="text-black bg-white focus:outline-none"
-//                             placeholder="First Name"
-//                         />
-//                     </label>
-
-//                     {/* Last Name Input */}
-//                     <label className="outline outline-1 pl-8 pr-20 py-3 rounded-lg bg-white">
-//                         <input
-//                             type="text"
-//                             value={lastName}
-//                             onChange={(e) => {
-//                                 setLastName(e.target.value);
-//                             }}
-//                             className="text-black bg-white focus:outline-none"
-//                             placeholder="Last Name"
-//                         />
-//                     </label>
-
-//                     {error && <p className="text-red-500 text-center">{error}</p>}
-
-//                     {/* Mobile-only continue button */}
-//                     <div className="block md:hidden lg:hidden text-center mt-8">
-//                         <button
-//                             onClick={handleContinue}
-//                             disabled={loading}
-//                             className="px-10 py-2 rounded-lg text-lg bg-[#18328D] text-white hover:bg-[#18328D] transition-all"
-//                         >
-//                             Continue
-//                         </button>
-//                     </div>
-//                 </div>
-//             </section>
-
-//             {/* Desktop/tablet layout with logo and continue button */}
-//             <div className="hidden md:flex left-72 right-72 bottom-44 items-center justify-around p-6">
-//                 <Image
-//                     src="/webclickericon.svg"
-//                     alt="logo"
-//                     width={112} // 28 * 4 = 112px
-//                     height={112}
-//                     className="w-28 h-28"
-//                     priority
-//                 />
-//                 <button
-//                     onClick={handleContinue}
-//                     disabled={loading}
-//                     className="px-10 py-2 rounded-lg text-lg bg-[#18328D] text-white hover:bg-[#18328D] transition-all"
-//                 >
-//                     Continue
-//                 </button>
-//             </div>
-//         </main>
-//     );
-// }
